scripts/requirements.txt
gTTS==2.5.1
moviepy==1.0.3
numpy==1.26.4
Pillow==11.0.0
imageio-ffmpeg==0.4.9
